+++ Netkan/Extensions/JObjectExtensions.cs
@@ -30,6 +30,26 @@ namespace CKAN.NetKAN.Extensions
         }
         
         /// <summary>
+        /// Write a property to a <see cref="JObject"/> only if it does not already exist and the value is not null.
+        /// The value is generated on-demand by calling tokenCallback, only when we need it.
+        /// </summary>
+        /// <param name="jobject">The <see cref="JObject"/> to write to</param>
+        /// <param name="propertyName">The name of the property to write</param>
+        /// <param name="tokenCallback">Function to generate value of the property to write if it does not exist</param>
+        public static void SafeAdd(this JObject jobject, string propertyName, Func<JToken> tokenCallback)
+        {
+            if (String.IsNullOrWhiteSpace(propertyName))
+            {
+                Log.Warn("Asked to set a property named null on a JSON object!");
+                return;
+            }
+            if (jobject[propertyName] == null)
+            {
+                jobject.SafeAdd(propertyName, tokenCallback());
+            }
+        }
+        
+        /// <summary>
         /// Merge an object's properties into one of our child objects
         /// E.g., the "resources" object should accumulate values from all levels
         /// </summary>